requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.24.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0

# Video processing
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
imageio>=2.33.0
imageio-ffmpeg>=0.4.9

# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0
python-dotenv>=1.0.0
pathlib>=1.0.1
logging>=0.5.1.2
playwright>=1.40.0

# Analysis
scipy>=1.11.0
scikit-learn>=1.3.0

# Video editing
moviepy>=1.0.3

# Pose estimation
tensorflow>=2.12.0
tensorflow-hub>=0.14.0

# Added from the code block
gradio>=4.0.0 